buy_stuff_mcp/requirements.txt
mcp
google-adk
python-dotenv
supabase
resend